Fix exe/terminalwire to run with rbenv
Why did I have to use this instead of terminalwire-exec?

Because rbenv executes the contents of the file via Kernal#load, which
reads the file and thinks its Ruby.

This is a bug in rbenv: ideally Rbenv reads the shebang and if its Ruby,
it calls Kernal#load. If its not Ruby, it would call Kernel#system or not
generate a shim at all.

Even if I fix the bug, it doesn't matter because people have to run this
with older versions of rbenv and I can't require people to upgrade. Instead
I just switched the binary back to Ruby.

terminalwire-exec will still work for other shims! Just not for terminalwire
binary.