Break apart the Thor generator into two files: MainTerminal and ApplicationTerminal
This makes it easier for people who are working with Thor for the first
time to add SubCommands into their terminals. I was on the fence about
this change until I watched another user build out a Terminalwire CLI
app and watched them quickly need to integrate sub-commands. This approach
worked very quickly.