Add experimental channel helper
This makes it possible for people to send context from the terminal to
other processes, like a browser, then wait for the process to finish
from the terminal before continuing.

This is useful when a browser needs to be opened to finish something before
allowing the terminal to continue, like authentication.